Ensure test resets environment variable.
