+++ src/test/java/de/zalando/ep/zalenium/registry/ZaleniumRegistryTest.java
@@ -73,7 +73,7 @@ public class ZaleniumRegistryTest {
 
         try {
             registry.add(p1);
-            
+
             await().pollInterval(Duration.FIVE_HUNDRED_MILLISECONDS).atMost(Duration.TWO_SECONDS).until(() -> registry.getAllProxies().size() == 1);
 
             RequestHandler newSessionRequest = TestUtils.createNewSessionHandler(registry, requestedCapability);
@@ -88,20 +88,5 @@ public class ZaleniumRegistryTest {
         } finally {
             registry.stop();
         }
-    }
-
-    /*
-        Uncomment the two bottom lines to run Zalenium in development mode.
-        Useful for implementing new features or debugging issues.
-     */
-    @Test
-    public void runLocally() {
-        System.setProperty("runningLocally", "true");
-        GridHubConfiguration gridHubConfiguration = new GridHubConfiguration();
-        gridHubConfiguration.registry = ZaleniumRegistry.class.getCanonicalName();
-        gridHubConfiguration.port = 4445;
-        Hub hub = new Hub(gridHubConfiguration);
-        // hub.start();
-        // Thread.sleep(1000 * 60 * 60);
     }
 }
